+++ test/generator_tests/models_generator_test.dart
@@ -12,14 +12,12 @@ void main() {
     GeneratorOptions(
       inputFolder: '',
       outputFolder: '',
-      nullableFields: false,
     ),
   );
   final generator2 = SwaggerModelsGeneratorV2(
     GeneratorOptions(
       inputFolder: '',
       outputFolder: '',
-      nullableFields: false,
     ),
   );
 
@@ -292,7 +290,7 @@ void main() {
       const className = 'Animals';
       const propertyKey = 'Dog';
       const jsonKeyExpectedResult = "\t@JsonKey(name: 'Dog')\n";
-      const fieldExpectedResult = '\tfinal Pet dog';
+      const fieldExpectedResult = '\tfinal Pet? dog';
       final result = generator.generatePropertyContentBySchema(
         map,
         propertyName,
@@ -317,7 +315,7 @@ void main() {
 
       const className = 'Animals';
       const jsonKeyExpectedResult = "\t@JsonKey(name: 'Animals')\n";
-      const fieldExpectedResult = 'final Pet animals';
+      const fieldExpectedResult = 'final Pet? animals';
       final result = generator.generatePropertiesContent(
         SwaggerRoot.empty,
         map,
@@ -344,7 +342,7 @@ void main() {
 
       const className = 'Animals';
       const jsonKeyExpectedResult = "\t@JsonKey(name: 'Animals')\n";
-      const fieldExpectedResult = 'final Pet animals';
+      const fieldExpectedResult = 'final Pet? animals';
       final result = generator.generatePropertiesContent(
         SwaggerRoot.empty,
         map,
@@ -369,7 +367,7 @@ void main() {
 
       const className = 'Animals';
       const jsonKeyExpectedResult = "\t@JsonKey(name: 'animals')\n";
-      const fieldExpectedResult = 'final Pet animals';
+      const fieldExpectedResult = 'final Pet? animals';
       final result = generator.generatePropertiesContent(
         SwaggerRoot.empty,
         map,
@@ -394,7 +392,7 @@ void main() {
 
       const className = 'Animals';
       const jsonKeyExpectedResult = "\t@JsonKey(name: '\\\$with')\n";
-      const fieldExpectedResult = 'final Pet \$with';
+      const fieldExpectedResult = 'final Pet? \$with';
       final result = generator.generatePropertiesContent(
         SwaggerRoot.empty,
         map,
@@ -422,7 +420,7 @@ void main() {
       const jsonKeyExpectedResult =
           "@JsonKey(name: 'Dog', defaultValue: <Object>[])";
 
-      const propertyExpectedResult = 'final List<Object> dog';
+      const propertyExpectedResult = 'final List<Object>? dog';
       final result = generator.generateListPropertyContent(
         propertyName,
         propertyKey,
@@ -482,7 +480,7 @@ void main() {
         {},
       );
 
-      expect(result, contains('final List<TestOriginalRef> dog;'));
+      expect(result, contains('final List<TestOriginalRef>? dog;'));
     });
 
     test('Should return List<Object> by ref', () {
@@ -505,7 +503,7 @@ void main() {
         {},
       );
 
-      expect(result, contains('final List<TestObject> dog;'));
+      expect(result, contains('final List<TestObject>? dog;'));
     });
   });